+++ client/js/table.js
@@ -664,4 +664,3 @@ function modalTable(params) {
 
     $('#tableModal').modal('show');
 }
-