+++ server/backends/plog/clickhouse/clickhouse.php
@@ -464,7 +464,7 @@
 
                 return $this->clickhouse->select($query);
             }
-            
+
             /**
              * @inheritDoc
              */