+++ server/hw/ip/domophone/akuvox/akuvox.php
@@ -592,4 +592,23 @@ abstract class akuvox extends domophone
     {
         $this->setConfigParams(['Config.DoorSetting.RFCARDDISPLAY.RfidDisplayMode' => '4']); // 8HR
     }
+
+    /**
+     * Sets the inversion mode for relays.
+     *
+     * @param bool $invertA Whether relay A should operate in inverted mode.
+     * @param bool $invertB Whether relay B should operate in inverted mode.
+     * @return void
+     */
+    protected function setRelayInversion(bool $invertA = false, bool $invertB = false): void
+    {
+        $this->apiCall('', 'POST', [
+            'target' => 'relay',
+            'action' => 'set',
+            'data' => [
+                'Config.DoorSetting.RELAY.RelayAType' => $invertA ? '1' : '0',
+                'Config.DoorSetting.RELAY.RelayBType' => $invertB ? '1' : '0',
+            ],
+        ]);
+    }
 }